+++ resources/views/components/table.blade.php
@@ -1,11 +1,13 @@
 <div class="overflow-x-auto">
     <div class="align-middle inline-block min-w-full">
-        <div class="shadow overflow-hidden border-b border-gray-200 dark:border-gray-500 sm:rounded-b-sm">
-            <table class="min-w-full divide-y divide-gray-200 dark:divide-gray-500">
-                <thead class="bg-gray-100 dark:bg-slate-700 select-none">
-                    {{ $head }}
+        <div class="overflow-hidden overflow-x-auto shadow-md sm:rounded-lg">
+            <table class="w-full text-sm text-left text-gray-500 dark:text-gray-400">
+                <thead class="text-xs text-gray-700 uppercase bg-gray-100 dark:bg-gray-700 dark:text-gray-400">
+                    <tr>
+                        {{ $head }}
+                    </tr>
                 </thead>
-                <tbody class="bg-white dark:bg-slate-600 divide-y divide-gray-200 dark:divide-gray-600">
+                <tbody>
                     {{ $body }}
                 </tbody>
             </table>