+++ resources/views/components/table/cell.blade.php
@@ -1,5 +1,6 @@
-<td {{ $attributes->except('class') }} @class([ 'px-4 py-4 whitespace-nowrap' , 'hidden lg:table-cell'=> $mobileOnly,
-    'w-full' => $fullWidth])>
+<td {{ $attributes->except('class') }} {{ $attributes->class(['px-4 py-4 whitespace-nowrap text-sm', 'hidden
+    lg:table-cell' => $mobileOnly, 'w-full' =>
+    $fullWidth]) }}>
     <div class="">
         {{ $slot }}
     </div>