+++ resources/views/components/table/row.blade.php
@@ -1,3 +1,4 @@
-<tr class="text-xs">
+<tr
+    class="bg-white last:border-0 border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600">
     {{ $slot }}
 </tr>